Finish moving logic
